Mostly made a bunch of comments, mainly in Bubbles
So many comments. Sorry about that.
Also, I don't really like having literally all of the buttons in
Bubbles. You said the code didn't work anywhere else, but if we had time
(which we don't, so like in a mythical world), I would want the card
buttons to be in a separate class or something. I realize we just really
need to realize on getting everything to work though.